[pdc] [msvc-2015] Change a variable type to prevent 32-bit MSVC-2015 "internal error"
[...]\packet-pdc.c(205) : fatal error C1001: An internal error has occurred in the compiler.

(compiler file 'f:\dd\vctools\compiler\utc\src\p2\main.c', line 246)
 To work around this problem, try simplifying or changing the program near the locations listed above.
Please choose the Technical Support command on the Visual C++
 Help menu, or open the Technical Support help file for more information

INTERNAL COMPILER ERROR in 'C:\Program Files\Microsoft Visual Studio 14.0\VC\BIN\cl.EXE'
    Please choose the Technical Support command on the Visual C++
    Help menu, or open the Technical Support help file for more information

Reviewed-on: https://code.wireshark.org/review/10005

+++ epan/dissectors/packet-pdc.c
@@ -147,7 +147,7 @@ static const value_string valstr_drmpdu_reason[] = {
 /* start of functions here */
 static int dissect_simpdu(tvbuff_t *tvb, proto_tree *tree, guint16 offset, guint8 lenIndicator)
 {
-	guint16	    bytesProcessed;
+	gint	    bytesProcessed;
 	guint8	    paramCode;
 	proto_item *simpduItem;
 	proto_tree *simpduVarTree;
